% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalhabml-package.R
\docType{package}
\name{globalhabml}
\alias{globalhabml}
\title{globalhabml: An introduction to machine learning for the Global HAB Workshop May 9-13, 2022.}
\description{
Machine learning tutorial examples and tools
}
